requirements.txt
click==6.7
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mod-wsgi==4.6.4
mysqlclient==1.3.12
passlib==1.7.1
pkg-resources==0.0.0
PyMySQL==0.8.1
PyMysqlDB==0.0.2
PyYAML==3.12
redis==2.10.6
SQLAlchemy==1.2.8
Werkzeug==0.14.1
